Fix ValidatePlugins test performance
